精简测试 Demo
