+++ app/src/main/java/me/yifeiyuan/flapdev/components/TestConfig.kt
@@ -1,8 +1,11 @@
 package me.yifeiyuan.flapdev.components
 
+import android.util.Log
+import android.widget.ImageView
 import me.yifeiyuan.flap.differ.IDiffer
 import me.yifeiyuan.flap.dsl.adapterDelegate
 import me.yifeiyuan.flap.ext.bindTextView
+import me.yifeiyuan.flap.ext.bindView
 import me.yifeiyuan.flapdev.R
 
 /**
