--- app/src/main/java/me/yifeiyuan/flapdev/components/base/BaseFlapComponent.java
@@ -1,18 +1,0 @@
-package me.yifeiyuan.flapdev.components.base;
-
-import android.view.View;
-
-import me.yifeiyuan.flap.Component;
-
-/**
- * Created by 程序亦非猿 on 2018/12/29.
- *
- * 可选
- */
-public abstract class BaseFlapComponent<T extends BaseModel> extends Component<T> {
-
-    public BaseFlapComponent(final View itemView) {
-        super(itemView);
-    }
-
-}
