--- app/src/main/java/me/yifeiyuan/flapdev/components/base/BaseModel.java
@@ -1,9 +1,0 @@
-package me.yifeiyuan.flapdev.components.base;
-
-/**
- * Created by 程序亦非猿 on 2019/1/29.
- *
- * 可选
- */
-public class BaseModel {
-}
