--- app/src/main/java/me/yifeiyuan/flapdev/components/generictest/GenericModel.java
@@ -1,9 +1,0 @@
-package me.yifeiyuan.flapdev.components.generictest;
-
-import me.yifeiyuan.flapdev.components.base.BaseModel;
-
-/**
- * Created by 程序亦非猿 on 2019/1/29.
- */
-public class GenericModel extends BaseModel {
-}
