+++ ktmodule/src/main/res/layout/flap_item_vb.xml
