--- othermodule/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest package="me.yifeiyuan.ktx.foundation.othermodule">
-
-    /
-</manifest>
