--- othermodule/src/main/java/me/yifeiyuan/ktx/foundation/othermodule/JavaModuleComponent.java
@@ -1,23 +1,0 @@
-package me.yifeiyuan.ktx.foundation.othermodule;
-
-import android.view.View;
-
-import androidx.annotation.NonNull;
-
-import me.yifeiyuan.flap.Component;
-import me.yifeiyuan.flap.annotations.Delegate;
-
-/**
- * Created by 程序亦非猿 on 2020/9/21.
- */
-@Delegate(layoutName = "java_module_component")
-public class JavaModuleComponent extends Component<JavaModuleModel> {
-    public JavaModuleComponent(@NonNull View itemView) {
-        super(itemView);
-    }
-
-    @Override
-    public void onBind(JavaModuleModel model) {
-
-    }
-}
