--- othermodule/src/main/res/layout/java_module_component.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-android:layout_width="match_parent"
-android:layout_height="50dp"
-android:layout_marginBottom="4dp"
-android:background="#AB89E8"
-android:orientation="horizontal">
-
-<TextView
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center"
-    android:paddingLeft="16dp"
-
-    android:text="Java Component In A Submodule"
-    android:textColor="@android:color/white"
-    android:textSize="16sp" />
-
-</LinearLayout>
